updated abstract_turtle lib
