+++ static/python/overrides/abstract_turtle/__init__.py
@@ -1,0 +1,3 @@
+from .turtle_class import Turtle
+from .canvas import Canvas
+from .logging_canvas import LoggingCanvas
