+++ static/python/overrides/abstract_turtle/pillow_canvas.py
@@ -1,3 +1,4 @@
+
 from math import sin, cos, pi
 from PIL import Image, ImageDraw
 import numpy as np
@@ -10,12 +11,12 @@ class PillowCanvas(Canvas):
     """
     Draw the given results on a pillow canvas.
     """
-
     def __init__(self, width, height):
         super().__init__(width, height)
-        self.image = Image.new("RGBA", (width, height))
+        self.image = Image.new('RGBA', (width, height))
         self.background_color = Color(255, 255, 255)
         self.draw = ImageDraw.Draw(self.image)
+        self.turtle = None
 
     def tr_pos(self, pos):
         x, y = pos.x, pos.y
@@ -26,14 +27,12 @@ class PillowCanvas(Canvas):
         return color.red, color.green, color.blue, 255
 
     def draw_rectangular_line(self, start, end, color, width):
-        self.draw.line(
-            [self.tr_pos(start), self.tr_pos(end)], self.tr_color(color), width
-        )
+        self.draw.line([self.tr_pos(start), self.tr_pos(end)], self.tr_color(color), width)
 
     def draw_circle(self, center, radius, color, width, is_filled):
         x, y = self.tr_pos(center)
-        left_up = (x - radius, y - radius)
-        right_down = (x + radius, y + radius)
+        left_up = (x-radius, y-radius)
+        right_down = (x+radius, y+radius)
         box = [left_up, right_down]
         if is_filled:
             self.draw.ellipse(box, fill=self.tr_color(color))
@@ -42,7 +41,8 @@ class PillowCanvas(Canvas):
 
     def fill_polygon(self, points, color):
         self.draw.polygon(
-            [self.tr_pos(point) for point in points], fill=self.tr_color(color)
+            [self.tr_pos(point) for point in points],
+            fill=self.tr_color(color)
         )
 
     def set_bgcolor(self, color):
@@ -52,12 +52,20 @@ class PillowCanvas(Canvas):
         self.draw.rectangle((0, 0, self.width, self.height), fill=(0, 0, 0, 0))
 
     def export(self):
-        data = np.array(self.image)
+        image = self.image.copy()
+        if self.turtle is not None:
+            ImageDraw.Draw(image).polygon(
+                [self.tr_pos(point) for point in self.turtle.points],
+                fill=self.tr_color(Color.of(0, 0, 0))
+            )
+        data = np.array(image)
         assert len(data.shape) == 3 and data.shape[-1] == 4
-        transparents = data[:, :, -1] == 0
+        transparents = data[:,:,-1] == 0
         data[transparents] = self.tr_color(self.background_color)
         return Image.fromarray(data)
 
+    def update_turtle(self, turtle):
+        self.turtle = turtle
 
 def circle(draw, cx, cy, r, fill, width=1, segments=100):
     # based on https://gist.github.com/skion/9259926
@@ -65,7 +73,7 @@ def circle(draw, cx, cy, r, fill, width=1, segments=100):
     dl = r * da
 
     for i in range(segments):
-        a = (i + 0.5) * da
+        a = (i+0.5) * da
 
         x = cx + r * cos(a)
         y = cy + r * sin(a)
@@ -73,4 +81,4 @@ def circle(draw, cx, cy, r, fill, width=1, segments=100):
         dx = -sin(a) * dl
         dy = cos(a) * dl
 
-        draw.line([(x - dx, y - dy), (x + dx, y + dy)], fill=fill, width=width)
+        draw.line([(x-dx, y-dy), (x+dx, y+dy)], fill=fill, width=width)
